requirements.txt
pandas
dotenv
streamlit
google-generativeai